rename navBar folder to navigation
Try to force vercel build cache clear
