+++ app/layout.tsx
@@ -12,7 +12,7 @@ import {
 import Sidebar from "./components/sidebar/sidebar";
 import { auth } from "../auth";
 import LoginPage from "pages/login/page";
-import { NavBar } from "components/navBar/NavBar";
+import { NavBar } from "components/navigation/NavBar";
 
 const poppins = Poppins({ weight: "400", style: "normal", subsets: ["latin"] });
 // trigger rebuild
